Step 3: Final Validation - Compile: SUCCESS | Tests: 1/1 passed
Upgraded PostgreSQL JDBC driver to 42.7.8 to address high-severity CVEs and fixed Action entity constructor type mismatch.

+++ demo/src/main/java/com/example/demo/entities/Action.java
@@ -41,7 +41,7 @@ public class Action {
     public Action() {
     }
 
-    public Action(Integer id, String title, String description, String location, Double goalAmount, Organisation organisation, List<Donation> donations) {
+    public Action(Long id, String title, String description, String location, Double goalAmount, Organisation organisation, List<Donation> donations) {
         this.id = id;
         this.title = title;
         this.description = description;
